refactor(sdk/links): move functions to their own files
