+++ packages/sdk/src/scopes/utilities/links/exports.ts
@@ -1,6 +1,3 @@
-export {
-  openLink,
-  openTelegramLink,
-  shareURL,
-  type OpenLinkOptions,
-} from './links.js';
+export { openLink, type OpenLinkOptions } from './openLink.js';
+export { openTelegramLink } from './openTelegramLink.js';
+export { shareURL } from './shareURL.js';
